# Remove unnecessary draining of HTTP request body from benchmark