It is not necessary to close HTTP request bodies, this is only necessary for HTTP response bodies. This has a negligible effect on the benchmark:

> name                  old time/op    new time/op    delta
> DeliveryNoReply-16       114µs ± 0%     123µs ± 0%   ~     (p=1.000 n=1+1)
> DeliveryWithReply-16     253µs ± 0%     240µs ± 0%   ~     (p=1.000 n=1+1)
> 
> name                  old alloc/op   new alloc/op   delta
> DeliveryNoReply-16      10.1kB ± 0%    10.2kB ± 0%   ~     (p=1.000 n=1+1)
> DeliveryWithReply-16    20.2kB ± 0%    20.2kB ± 0%   ~     (p=1.000 n=1+1)
> 
> name                  old allocs/op  new allocs/op  delta
> DeliveryNoReply-16         143 ± 0%       143 ± 0%   ~     (all equal)
> DeliveryWithReply-16       301 ± 0%       301 ± 0%   ~     (all equal)